requirements.txt
astroid==2.2.5
autopep8==1.4.4
certifi==2019.6.16
chardet==3.0.4
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.3
django-allauth==0.39.1
django-ckeditor==5.7.1
django-cors-headers==3.0.2
django-cuser==2017.3.16
django-dropbox-storage==0.4.1
django-heroku==0.3.1
django-js-asset==1.2.2
django-modeltranslation==0.13.4
django-rest-auth==0.9.5
django-storages==1.7.2
django-tinymce==2.8.0
djangorestframework==3.9.4
dropbox==9.4.0
gunicorn==19.9.0
idna==2.8
isort==4.3.21
jsonfield==2.0.2
lazy-object-proxy==1.4.1
mccabe==0.6.1
numpy==1.17.1
oauthlib==3.0.1
pep8==1.7.1
Pillow==6.1.0
psycopg2==2.8.3
pudb==2019.1
pycodestyle==2.5.0
pyDOE==0.3.8
Pygments==2.4.2
pylint==2.3.1
python3-openid==3.1.0
pytz==2019.1
requests==2.22.0
requests-oauthlib==1.2.0
scipy==1.3.1
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
urllib3==1.25.3
urwid==2.0.1
whitenoise==4.1.3
wrapt==1.11.2
